@@ -1,14 +1,11 @@
 package com.sahnisemanyazilim.ezanisaat;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.Looper;
-import android.support.v7.widget.Toolbar;
+
+import androidx.appcompat.widget.Toolbar;
 import android.view.MenuItem;
 
 import com.kemalettinsargin.mylib.MyFragmentActivity;
-import com.kemalettinsargin.mylib.Util;
 
 
 public class HakkindaActivity extends MyFragmentActivity {
